# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Contractor Estimate Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice with details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceRequest"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List all materials
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: supplierId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create a material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaterialRequest"
      responses:
        "201":
          description: Created material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    put:
      operationId: updateMaterial
      tags: [materials]
      summary: Update a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaterialRequest"
      responses:
        "200":
          description: Updated material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /materials/{id}/price-history:
    get:
      operationId: getMaterialPriceHistory
      tags: [materials]
      summary: Get price history for a material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceHistory"
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
...
    LineItem:
      type: object
      properties:
        id:
          type: integer
        estimateId:
          type: integer
        type:
          type: string
          description: "labor | material | subcontractor | other"
        category:
          type: string
          nullable: true
        description:
          type: string
        quantity:
          type: number
        unit:
          type: string
          nullable: true
        unitCost:
          type: number
        total:
          type: number
        materialId:
          type: integer
          nullable: true
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required:
        - id
        - estimateId
        - type
        - description
        - quantity
        - unitCost
        - total
        - sortOrder
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        estimateId:
          type: integer
          nullable: true
        clientId:
          type: integer
          nullable: true
        clientName:
          type: string
          nullable: true
        projectName:
          type: string
        status:
          type: string
          description: "draft | sent | paid | overdue | cancelled"
        subtotal:
          type: number
        markupAmount:
          type: number
        taxAmount:
          type: number
        total:
          type: number
        amountPaid:
          type: number
        amountDue:
          type: number
        dueDate:
          type: string
          nullable: true
        paidAt:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - invoiceNumber
        - projectName
        - status
        - subtotal
        - markupAmount
        - taxAmount
        - total
        - amountPaid
        - amountDue
        - createdAt
    InvoiceDetail:
      allOf:
        - $ref: "#/components/schemas/Invoice"
        - type: object
          properties:
            lineItems:
              type: array
              items:
                $ref: "#/components/schemas/LineItem"
            client:
              $ref: "#/components/schemas/Client"
              nullable: true
          required:
            - lineItems
    CreateInvoiceRequest:
      type: object
      properties:
        estimateId:
          type: integer
          nullable: true
        clientId:
          type: integer
          nullable: true
        projectName:
          type: string
        status:
          type: string
          default: draft
        dueDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - projectName
    UpdateInvoiceRequest:
      type: object
      properties:
        status:
          type: string
        dueDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        amountPaid:
          type: number
        paidAt:
          type: string
          nullable: true
    Material:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        currentPrice:
          type: number
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        sku:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        lastUpdated:
          type: string
      required:
        - id
        - name
        - currentPrice
        - lastUpdated
    CreateMaterialRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        currentPrice:
          type: number
        supplierId:
          type: integer
          nullable: true
        sku:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
      required:
        - name
        - currentPrice
    PriceHistory:
      type: object
      properties:
        id:
          type: integer
        materialId:
          type: integer
        price:
          type: number
        recordedAt:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - id
        - materialId
        - price
        - recordedAt
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        website:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - id
        - name
    CreateSupplierRequest:
      type: object
      properties:
        name:
          type: string
        website:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
...