# lib/api-zod/src/generated/types/createSupplierRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Contractor Estimate Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSupplierRequest {
  name: string;
  website?: string | null;
  phone?: string | null;
  notes?: string | null;
}