# lib/api-zod/src/generated/types/invoiceDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Contractor Estimate Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from "./client";
import type { Invoice } from "./invoice";
import type { LineItem } from "./lineItem";

export type InvoiceDetail = Invoice & {
  lineItems: LineItem[];
  client?: Client | null;
};