# lib/api-zod/src/generated/types/priceHistory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Contractor Estimate Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PriceHistory {
  id: number;
  materialId: number;
  price: number;
  recordedAt: string;
  notes?: string | null;
}