# lib/api-zod/src/generated/types/updateInvoiceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Contractor Estimate Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateInvoiceRequest {
  status?: string;
  dueDate?: string | null;
  notes?: string | null;
  amountPaid?: number;
  paidAt?: string | null;
}